modal works (finally)
