+++ vue-frontend/src/components/Modal.vue
@@ -1,10 +1,11 @@
 <script setup>
-  import { ref } from 'vue';
+  import { ref, watchEffect } from 'vue';
 
-  const modalEnabled = defineModel();
+  const modalEnabled = defineModel()
+  function closeModal() { modalEnabled.value = !modalEnabled.value }
 </script>
 <template>
-  <span @click="modalEnabled = !modalEnabled" :class="{ 'fixed flex justify-center z-50 top-0 left-0 h-full w-full bg-black/50': modalEnabled, 'hidden': !modalEnabled }">
+  <span @click.self="closeModal" @keyup.esc="closeModal" :class="{ 'fixed flex justify-center z-10 top-0 left-0 h-full w-full bg-black/50': modalEnabled, 'hidden': !modalEnabled }">
     <slot />
   </span>
 </template>
