+++ vue-frontend/src/components/TopBar.vue
@@ -14,44 +14,11 @@
   const userRole = route.params.userRole
 
   const enableButton = defineModel('enableButton')
-  const modalEnabled = ref(false)
+  const emit = defineEmits(['modalEnabled'])
 
   const classroomCode = ref('')
   const classroomName = ref('');
   const classroomDesc = ref('');
-
-  async function joinClassroom() {
-    const join = await axios.put(API + 'student/' + userId + '/classroom-join/' + classroomCode.value)
-    console.log(join.data)
-
-    if (join.data.status === 'success') {
-      alert('Joined classroom successfully, redirecting to classroom...')
-      let classroomId = join.data.classroom_id
-      router.push({ path: '/classroom/' + classroomId }) 
-    }
-    else if (join.data.status === 'already joined') {
-      alert(join.data.message)
-    }
-    else if (join.data.status === 'classroom not found') {
-      alert(join.data.message) 
-    }
-  }
-
-  async function createClassroom() {
-    const classroomData = {
-      classroom_name: classroomName.value,
-      classroom_desc: classroomDesc.value,
-    }
-    try {
-      const res = await axios.post(API + 'classroom/create', classroomData)
-      if (res.data.status === 'success') {
-        alert('Classroom created successfully, redirecting...')
-        router.push(res.data.classroom_id) 
-      }
-    } catch (err) {
-      console.log(err)
-    }
-  }
 
   onMounted(() => {
     console.log(route.meta)
@@ -64,42 +31,13 @@
       <RouterLink :to="{ name: 'userHome' }"> E-Quizz </RouterLink>
 
       <!--Create/join classroom button-->
-      <button @click="modalEnabled = !modalEnabled" v-if="userRole === 'student' && pageName.includes('Home') && enableButton" class="ml-2 bg-blue-300 hover:bg-blue-600 duration-150 ease-in hover:text-white text-black rounded-md px-2">
+      <button @click="emit('modalEnabled')" v-if="userRole === 'student' && enableButton" class="ml-2 bg-blue-300 hover:bg-blue-600 duration-150 ease-in hover:text-white text-black rounded-md px-2">
         + Join a classroom 
       </button>
 
-      <button @click="createClassroomModal = !createClassroomModal" v-else-if="userRole === 'teacher' && pageName.includes('Home') && enableButton"  to="classroom/create" class="ml-2 bg-blue-300 hover:bg-blue-600 duration-150 ease-in hover:text-white text-black rounded-md px-2"> 
+      <button @click="emit('modalEnabled')" v-else-if="userRole === 'teacher' && enableButton" class="ml-2 bg-blue-300 hover:bg-blue-600 duration-150 ease-in hover:text-white text-black rounded-md px-2"> 
         + Create a classroom
       </button>
-
-      <!-- Join Classroom Modal -->
-      <Modal v-if="userRole === 'student'" v-model="modalEnabled">
-        <div class="bg-gray-600 rounded-md w-3/4 h-fit text-2xl flex place-self-center">
-          <div class="flex flex-col justify-around my-4 mx-4">
-            <p>Enter classroom code</p>
-            <input type="text" placeholder="Classroom Code" v-model="classroomCode" class="rounded-md text-black w-3/4" />
-            <button @click="joinClassroom" class="w-1/4 rounded-md bg-blue-400 hover:bg-blue-600 justify-self-end float-right"> Join </button>
-          </div>
-        </div>
-      </Modal>
-      <!-- ---------- -->
-
-      <!-- Create Classroom Modal -->
-      <Modal v-else v-model="modalEnabled">
-        <div class="bg-gray-600 rounded-md w-3/4 h-fit flex place-self-center">
-          <div class="flex flex-col text-black">
-            <input type="text" v-model="classroomName" placeholder="Classroom Name" class="border-2 border-gray-300 rounded-md p-2 m-2">
-            <input type="text" v-model="classroomDesc" placeholder="Description" class="border-2 border-gray-300 rounded-md p-2 m-2">
-          </div>
-
-          <div class="float-right flex flex-row mx-2">
-            <button @click="createClassroom" class="bg-blue-400 rounded-md px-2 mx-2">Create Classroom</button>
-            <button>Cancel</button>
-          </div>
-        </div>
-      </Modal>
-      <!-- -------------- -->
-
     </div>
 
     <!-- Page Name -->
